Moved setAmpExt to the amp_auction file.
